Fix build issues in content/ for Android x64
BUG=346626

Review URL: https://codereview.chromium.org/182833003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@254136 0039d316-1c4b-4281-b951-d872f2087c98

+++ content/browser/renderer_host/render_widget_host_view_android.h
@@ -221,7 +221,7 @@ class RenderWidgetHostViewAndroid
   void LockResources();
   void UnlockResources();
 
-  int GetNativeImeAdapter();
+  long GetNativeImeAdapter();
 
   void WasResized();
