+++ content/public/android/java/src/org/chromium/content/browser/ContentSettings.java
@@ -40,7 +40,7 @@ public class ContentSettings {
      * @param nativeContentSettings the native instance that is going away.
      */
     @CalledByNative
-    private void onNativeContentSettingsDestroyed(int nativeContentSettings) {
+    private void onNativeContentSettingsDestroyed(long nativeContentSettings) {
         assert mNativeContentSettings == nativeContentSettings;
         mNativeContentSettings = 0;
     }
